requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
tldextract>=5.1.0
tqdm>=4.66.0
aiohttp>=3.9.0
aiosocks>=0.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0